Image_steganography_by_sofiae/Image_steganography_by_sofiae/requirements.txt
blinker==1.8.2
click==8.1.7
docopt==0.6.2
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
Werkzeug==3.1.2
